Fixes before presentation

+++ backend/middleware/sharp-config.js
@@ -6,17 +6,21 @@ sharp.cache(false);
 
 const resizeImage = (req, res, next) => {
     try {
-        const ref = path.parse(req.file.originalname).name;
-        console.log(req.file);
-        sharp(req.file.path)
-            .resize({
-                width: 500,
-                height: 400
-            })
-            .webp({ quality: 100 })
-            .toFile('images/' + ref + ".webp")
-            .then(() => fs.unlink(req.file.path, (e) => console.log(e)));
-        next();
+        if (!req.file) {
+            return next();
+        } else {
+            const ref = path.parse(req.file.originalname).name;
+            console.log(req.file);
+            sharp(req.file.path)
+                .resize({
+                    width: 360,
+                    height: 570
+                })
+                .webp({ quality: 100 })
+                .toFile('images/' + ref + ".webp")
+                .then(() => fs.unlink(req.file.path, (e) => console.log(e)));
+            next();
+        }
     } catch (error) {
         console.log(error);
     }
